refactor: standardize collection initialization and add missing repository interface
- Add IProgramExerciseRepository interface with async CRUD methods
- Standardize all navigation collections to use null! pattern
- Add trailing newlines to all source files for consistency
- Update CLAUDE.md to reflect current project state

+++ WorkoutTracker.Core/Models/ProgramExercise.cs
@@ -7,7 +7,7 @@ namespace WorkoutTracker.Core.Models
         public int Order { get; set; }
         public int PlannedSets { get; set; }
         public int PlannedReps { get; set; }
-        public float PlannedWeight { get; set; }
+        public decimal PlannedWeight { get; set; }
         public int PlannedRestTimeSeconds { get; set; }
         public Program Program { get; set; } = null!;
         public Exercise Exercise { get; set; } = null!;
